+++ cypress/integration/01-view-panel.test.ts
@@ -1,0 +1,41 @@
+import { e2e } from '@grafana/e2e';
+import { TestIds } from '../../src/constants';
+
+/**
+ * Dashboard
+ */
+const json = require('../../provisioning/dashboards/panels.json');
+const testedPanel = json.panels[2];
+
+/**
+ * Selector
+ */
+const getTestIdSelector = (testId: string) => `[data-testid="${testId}"]`;
+
+/**
+ * Panel
+ */
+describe('Viewing a panel with an Image', () => {
+  beforeEach(() => {
+    e2e.flows.openDashboard({
+      uid: json.uid,
+    });
+  });
+
+  it('Should display an Image', () => {
+    const currentPanel = e2e.components.Panels.Panel.title(testedPanel.title);
+    currentPanel.should('be.visible');
+
+    /**
+     * Image
+     */
+    const image = currentPanel.find(getTestIdSelector(TestIds.panel.root));
+    image.should('be.visible');
+
+    /**
+     * Screenshot
+     */
+    image.screenshot(testedPanel.title);
+    e2e().compareScreenshots({ name: testedPanel.title, threshold: 0.05 });
+  });
+});