+++ src/components/ImagePanel/ImagePanel.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { css, cx } from '@emotion/css';
 import { FieldType, PanelProps } from '@grafana/data';
 import { Alert, useStyles2 } from '@grafana/ui';
-import { ImageSizeModes, ImageTypesSymbols, SupportedTypes } from '../../constants';
+import { ImageSizeModes, ImageTypesSymbols, SupportedTypes, TestIds } from '../../constants';
 import { getStyles } from '../../styles';
 import { base64toBlob } from '../../utils';
 
@@ -216,6 +216,7 @@ export const ImagePanel: React.FC<Props> = ({ options, data, width, height, repl
    */
   return (
     <div
+      data-testid={TestIds.panel.root}
       className={cx(
         styles.wrapper,
         css`